requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dill==0.4.0
distlib==0.4.0
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
isort==6.0.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mccabe==0.7.0
narwhals==2.0.1
nodeenv==1.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.3.0
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pylint==3.3.8
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.33.1
watchdog==6.0.0
websockets==15.0.1
